+++ public/javascripts/notifier.js
@@ -366,7 +366,7 @@ printStackTrace.implementation.prototype = {
             '</notifier>' +
             '<error>' +
                 '<class>{exception_class}</class>' +
-                '<message>{exception_message}</message>' +
+                '<message><![CDATA[{exception_message}]]></message>' +
                 '<backtrace>{backtrace_lines}</backtrace>' +
             '</error>' +
             '<request>' +
@@ -972,8 +972,8 @@ printStackTrace.implementation.prototype = {
                     //        '" number="' + matches[3] + '" />');
                     
                     backtrace.push({
-                        'function': matches[1],
-                        file: file,
+                        'function': Util.escape(matches[1]),
+                        file: Util.escape(file),
                         line: matches[3]
                     });
                 }