+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - [#343][] Fix the ical generation. ([@shingara][])
 - [#503][] Fix issue on where the service_url choose never use. ([@nfedyashev][])
 - [#506][] Fix issue on bitbucket issue tracker creation failed. ([@Gonzih][])
+- [#514][] Add CDATA in xml return by Javascript. ([@mildavw][])
+- [#517][] Javascript escape path from javascript Notifier. ([@roryf][])
+- [#518][] Fix issue when you try launch task errbit:db:update_update_problem_attrs. ([@shingara][])
 
 ## 0.1.0 - 2013-05-29
 
@@ -72,6 +75,8 @@
 [#503]: https://github.com/errbit/errbit/issues/503
 [#506]: https://github.com/errbit/errbit/issues/506
 [#508]: https://github.com/errbit/errbit/issues/508
+[#514]: https://github.com/errbit/errbit/issues/514
+[#517]: https://github.com/errbit/errbit/issues/517
 
 <!-- Contributor on Errbit Thanks to all of them -->
 
@@ -90,3 +95,5 @@
 [@parallel588]: https://github.com/parallel588
 [@Gonzih]: https://github.com/Gonzih
 [@boblail]: https://github.com/boblail
+[@roryf]: https://github.com/roryf
+[@mildavw]: https://github.com/mildavw